additional condition to delete user, added synchronized methods for repository updates

+++ src/main/java/com/pas/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.pas.controller;
 
 import com.pas.model.User;
+import com.pas.service.impl.RentService;
 import com.pas.service.impl.UserService;
 
 import java.util.List;
@@ -10,9 +11,10 @@ import java.util.stream.Collectors;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.validation.*;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
 import jakarta.ws.rs.*;
-import jakarta.ws.rs.Path;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
@@ -23,6 +25,8 @@ public class UserController {
 
     @Inject
     private UserService userService;
+    @Inject
+    private RentService rentService;
 
     @POST
     @Produces(MediaType.APPLICATION_JSON)
@@ -35,8 +39,13 @@ public class UserController {
             return Response.status(Response.Status.BAD_REQUEST).entity(errors).build();
         }
         user.setId(UUID.randomUUID());
-        userService.addClient(user);
-        return Response.status(Response.Status.CREATED).build();
+        user.setActive(false);
+        User createdUser = userService.addClient(user);
+        if (createdUser != null) {
+            return Response.status(Response.Status.CREATED).build();
+        } else {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
     }
 
     @GET
@@ -73,7 +82,9 @@ public class UserController {
     @Consumes(MediaType.TEXT_PLAIN)
     @Path("/{userId}")
     public void deleteClient(@PathParam("userId") UUID userId) {
-        userService.deleteClient(userService.getClientById(userId));
+        if (rentService.getRentByClient(getClient(userId)) == null) {
+            userService.deleteClient(userService.getClientById(userId));
+        }
     }
 
     @PUT
@@ -97,19 +108,39 @@ public class UserController {
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     @Path("/moderator")
-    public User saveModerator(User user) {
+    public Response saveModerator(User user) {
+        Set<ConstraintViolation<User>> violation = validator.validate(user);
+        List<String> errors = violation.stream().map(ConstraintViolation::getMessage).collect(Collectors.toList());
+        if (!violation.isEmpty()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(errors).build();
+        }
         user.setId(UUID.randomUUID());
-        userService.addModerator(user);
-        return user;
+        user.setActive(false);
+        User createdUser = userService.addModerator(user);
+        if (createdUser != null) {
+            return Response.status(Response.Status.CREATED).build();
+        } else {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
     }
 
     @POST
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     @Path("/admin")
-    public User saveAdmin(User user) {
+    public Response saveAdmin(User user) {
+        Set<ConstraintViolation<User>> violation = validator.validate(user);
+        List<String> errors = violation.stream().map(ConstraintViolation::getMessage).collect(Collectors.toList());
+        if (!violation.isEmpty()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(errors).build();
+        }
         user.setId(UUID.randomUUID());
-        userService.addAdmin(user);
-        return user;
+        user.setActive(false);
+        User createdUser = userService.addAdmin(user);
+        if (createdUser != null) {
+            return Response.status(Response.Status.CREATED).build();
+        } else {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
     }
 }
